+++ HISTORY.md
@@ -9,6 +9,7 @@
   avoid memory and timeout issues.
 - RNA-seq variant calling: allow annotation of output variants with vcfanno
 - Support hg38 builds with peddy QC
+- QC: support VerifyBamID2 for contamination detection
 - CWL: adjust defaults for align_split_size and nomap_split_targets to match
   different parallelization and overhead for these runs
 - CWL: support for Cromwell runner